requirements.txt
# featurewiz==0.2.3
filetype==1.0.13
hnswlib==0.6.2
matplotlib==3.5.1
moviepy==2.0.0.dev2
natsort==8.1.0
netCDF4==1.6.2
numpy==1.20.0
opencv_python==4.6.0.66
pandas==1.4.2
Pillow==9.3.0
projection_pursuit==1.0
pytube==12.1.0
scikit_learn==1.1.3
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==61.2.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
wget==3.2
xarray==2022.3.0
